third_party/blink/web_tests/wpt_internal/css/css-size-adjust/text-size-adjust-affects-line-height-px.html

<!DOCTYPE html>
<meta content="width=device-width,initial-scale=1.0,minimum-scale=1.0" name="viewport">
<meta name="assert" content="text-size-adjust should affect line-height specified in px.">
<link rel="help" href="https://drafts.csswg.org/css-size-adjust/#propdef-text-size-adjust">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script>
  function runTest() {
    // This should not be needed but is required to work around a bug in
    // Chrome's old autosizer due to computed style being updated during layout.
    // TODO(pdr): Remove this when b/340389272 launches.
    const forceLayout = document.body.offsetHeight;

    test(() => {
      assert_equals(getComputedStyle(test_12)['line-height'], '18px');
    }, 'Resolved line-height with specified 12px line-height and text-size-adjust: 150%');

    test(() => {
      assert_equals(test_12.getBoundingClientRect().height, 18);
    }, 'Actual line-height with specified 12px line-height and text-size-adjust: 150%');

    test(() => {
      assert_equals(getComputedStyle(test_14)['line-height'], '21px');
    }, 'Resolved line-height with specified 14px line-height and text-size-adjust: 150%');

    test(() => {
      assert_equals(test_14.getBoundingClientRect().height, 21);
    }, 'Actual line-height with specified 14px line-height and text-size-adjust: 150%');
  }
</script>
<body onload="runTest()">
  <div id="test_12" style="line-height: 12px; text-size-adjust: 150%;">test</div>
  <div id="test_14" style="line-height: 14px; text-size-adjust: 150%;">test</div>
</body>
